% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareNA.R
\name{compareNA}
\alias{compareNA}
\title{compareNA}
\usage{
compareNA(v1, v2)
}
\arguments{
\item{v1}{first thing, can be NA}

\item{v2}{second thing, can be NA}
}
\description{
\code{compareNA} taken from R cookbook -- enables comparison of two things (strings/numbers/whatever) that may or may not contain NA. Will return TRUE if both  are equal, including if both are NA.
}
\seealso{
\link[ http://www.cookbook-r.com/Manipulating_data/Comparing_vectors_or_factors_with_NA/]
}
