% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults_grid.R
\name{extractResults_grid}
\alias{extractResults_grid}
\title{extractResults_grid
\code{extractResults_grid} custom wrapper for Mako simulation study to extract reference points nested SS3 simulations}
\usage{
extractResults_grid(
  rootdir,
  pattern = NA,
  writeTables = T,
  FleetName = c("S4_JPN_SS", "S7_JPN_GEO", NA)[1]
)
}
\arguments{
\item{rootdir}{root filepath where all subdirectories containing Report.sso are stored}

\item{pattern}{a string that specifically matches all directories with report files of interest.}

\item{writeTables}{logical. Should a csv of results be saved?}

\item{FleetName}{optional vector of fleet names for which data should be extracted; all means all}
}
\description{
extractResults_grid
\code{extractResults_grid} custom wrapper for Mako simulation study to extract reference points nested SS3 simulations
}
