% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getF.R
\name{getF}
\alias{getF}
\title{getF
Simple effort model from Carruthers et al. (2012) - function extracted from his package DLMtool in R}
\usage{
getF(nsim, Esd, nyears, dFmin, dFmax, bb, scale)
}
\arguments{
\item{nsim}{Number of simulated effort datasets}

\item{Esd}{optional; if you already generated a MQ csv using extractResults it will read from that, otherwise need pattern/subpattern}

\item{nyears}{the character column name of the biomass ref point; can rename to change y lab}

\item{dFmin}{the character column name of the fishing ref point}

\item{dFmax}{1st order directory character}

\item{bb}{2nd order directory character}

\item{scale}{scaling for median F}
}
\description{
getF
Simple effort model from Carruthers et al. (2012) - function extracted from his package DLMtool in R
}
