% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMVNhes.R
\name{getMVNhes}
\alias{getMVNhes}
\title{get_mvnhes
Function to generate mvn from var and corr of log (F/Fmsy) and log(SSB/SSBmsy) extracted from the hessian. Author: Felipe Carvalho}
\usage{
getMVNhes(data = out, pars = c("Bratio", "F"), yr = 2017,
  yr.cov = NULL, mc = 10000)
}
\arguments{
\item{rootdir}{home directory}

\item{mq_csv}{optional; if you already generated a MQ csv using extractResults it will read from that, otherwise need pattern/subpattern}

\item{axes.limits}{x and y maximums for kobe plot, default to 2}

\item{kobe.type}{if 'ISC', will have pale orange coloring in place of orange/yellow}

\item{b.name}{the character column name of the biomass ref point; can rename to change y lab}

\item{f.name}{the character column name of the fishing ref point}

\item{pattern}{1st order directory character}

\item{subpattern}{2nd order directory character}

\item{saveplot}{logical, do you want to save the plot}

\item{plotloc}{where to save plot}

\item{doLegend}{logical. do you want to write a legend (not advised for huge rep sets)}
}
\description{
get_mvnhes
Function to generate mvn from var and corr of log (F/Fmsy) and log(SSB/SSBmsy) extracted from the hessian. Author: Felipe Carvalho
}
