% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeasonFrate.R
\name{getSeasonFrate}
\alias{getSeasonFrate}
\title{getSeasonFrate
splits up F rates generated by getF into seasonal, fleetwise vectors for use in SS3.}
\usage{
getSeasonFrate(Fsim = Fsim, nAnnualFrate = nyears)
}
\arguments{
\item{Fsim}{vector returned from getF}

\item{nAnnualFrate}{number of years for which to generate an annual F rate.}
}
\description{
getSeasonFrate
splits up F rates generated by getF into seasonal, fleetwise vectors for use in SS3.
}
