% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShape.R
\name{getShape}
\alias{getShape}
\title{getShape
generate a beta random variable given a mean and sd. If the mean is 1 it will throw an error.}
\usage{
getShape(phis, rng = F, ngen = NA)
}
\arguments{
\item{phis}{list of length two containing mean, cv}

\item{rng}{logical. if T, this will randomly generate a number under your params. if falso it will return alpha and beta (shape and rate).}

\item{ngen}{numeric. number of random number(s) to generate.}
}
\description{
getShape
generate a beta random variable given a mean and sd. If the mean is 1 it will throw an error.
}
