% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getERDDAP.R
\name{getSST}
\alias{getSST}
\title{GetSST}
\usage{
getSST(
  df,
  lat = "LAT",
  lon = "LON",
  hitrans = T,
  meanloc = F,
  grouploc = "Site",
  info = "erdBAsstamday",
  dates = c("2012-07-16", "2013-09-16"),
  buffer = 2
)
}
\arguments{
\item{df}{a dataframe containing latitudes ("LAT") and longitude ("LONG") values, either aggregated by site or individual records}

\item{lat}{the column name with decimal degrees latitude}

\item{lon}{the column name with decimal degrees longitude}

\item{hitrans}{logical. If T, negative coordinates will be recentered around Hawaii so the Pacific ocean is whole. Note that this is required for use with many errdap datasets.}

\item{meanloc}{logical. If T, function will determine mean lat and long for each site (if many records per site).}

\item{grouploc}{If meanloc is T, provide grouping column (e.g. "site") for mean lat and long.}

\item{info}{Optionally select another griddap dataset; default is erdBAsstamday}

\item{dates}{Two dates in YYYY-MM-DD format, between which all SST values will be averaged.}

\item{buffer}{Number in decimal degrees that will be added to each location when searching for data}
}
\description{
\code{GetSST} Retrieves 2x2 degree aggregated SST data from the ERDDAP SERVER https://github.com/ropensci/rerddap
}
\seealso{
\code{\link[rerddap]{ed_search(query = "OCEANWATCH SST GOES-POES")}}
}
