% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWCGOP_len.R
\name{getWCGOP_len}
\alias{getWCGOP_len}
\title{getWCGOP_len
beta
a function to transpose data from the WCGOP into a usable form for SS.
returns a table that can be saved/copied into a stock synthesis file.
are in the traditional format as from the Observer program}
\usage{
getWCGOP_len(
  lcomps,
  Nsamp = round(runif(13, 113, 554)),
  fleet,
  fleetno = 1,
  lbin_lo = NA,
  lbin_hi = NA,
  lbin_inc = NA,
  month = 1,
  sex = 0,
  part,
  writeTable = F,
  writeloc = getwd()
)
}
\arguments{
\item{lcomps}{a data frame of length-comps from the WCGOP program}

\item{Nsamp}{a vector of # samples. Typically this is N_Unique_trips from the WCGOP data.}

\item{fleet}{a string or vector of strings matching values from column Gear}

\item{fleetno}{optional; fill in any number to the "FltSvy" column of the resultant table; defaults to 1}

\item{lbin_lo}{minimum bin}

\item{lbin_hi}{maximum bin}

\item{lbin_inc}{increments of bins}

\item{month}{optional; fill in a month into the month column of the resultant table}

\item{sex}{As in SS: # sex codes:  0=combined; 1=use female only; 2=use male only; 3=use both as joint sexxlength distribution}

\item{part}{# partition codes:  (0=combined; 1=discard; 2=retained) will be written to table}

\item{writeTable}{logical. if T, requires writeloc}

\item{writeloc}{a filepath to where you want the table saved.}
}
\description{
getWCGOP_len
beta
a function to transpose data from the WCGOP into a usable form for SS.
returns a table that can be saved/copied into a stock synthesis file.
are in the traditional format as from the Observer program
}
