% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDecTable.R
\name{makeDecTable}
\alias{makeDecTable}
\title{makeDecTable}
\usage{
makeDecTable(rootdir, state_names = c("low", "base", "high"),
  catch_names = c("constant", "ABC", "upper"), years, baseCatch,
  row1Catch, row3Catch, round = T, writeTable = F,
  writeloc = getwd())
}
\arguments{
\item{rootdir}{a master folder with all nine executed models.}

\item{state_names}{three names referring to the states of nature, which must match names of the directories. defaults to low/base/high}

\item{catch_names}{three names referring to the catch scenarios, which must match names of the directories. defaults to constant/ABC/upper}

\item{years}{a vector of years for which you'd like output}

\item{baseCatch}{a vector of catch values used for the base case model(s) (center row) -- must match length(years)}

\item{row1Catch}{a vector of catch values used for the first row model(s) (typically constant or lower stream)}

\item{row3Catch}{a vector of catch values used for the first row model(s) (typically constant or lower stream)}

\item{round}{logical, defaults to TRUE}

\item{writeTable}{logical, defaults to FALSE}

\item{writeloc}{if \code{writeTable == T}, directory to write output}
}
\description{
\code{makeDecTable} given a directory or set of directories, will stitch together info for decision table of 3x3 dimensions
}
