% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestidx.R
\name{nestidx}
\alias{nestidx}
\title{Nested loop index}
\usage{
nestidx(vals = list(a, b, b.vector, c = 0, c.vector = 0))
}
\arguments{
\item{vals}{a list of up to length three with the current indices and length of each index vector}
}
\description{
\code{nestidx} returns the index value for a matrix populated by nested loops.
}
