% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCPUE_panel.R
\name{plotCPUE_panel}
\alias{plotCPUE_panel}
\title{plotCPUE_panel
\code{plotCPUE_panel} panel plot illustrating fitted CPUE from model(s) and observed/CI}
\usage{
plotCPUE_panel(
  cpue_results,
  saveplot = T,
  mods = NA,
  plotloc = NA,
  pdfrows = 4,
  pdfcols = 2
)
}
\arguments{
\item{cpue_results}{a data frame generated using extract_results()}

\item{saveplot}{logical. should plots be saved to model directory}

\item{mods}{vector of model names as they appear in the cpue_results CSV to use; NA will use all}

\item{plotloc}{folder to save plots; if NA, will save in location of cpue_results}
}
\description{
plotCPUE_panel
\code{plotCPUE_panel} panel plot illustrating fitted CPUE from model(s) and observed/CI
}
