% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotF.R
\name{plotF}
\alias{plotF}
\title{plotF}
\usage{
plotF(SPRseries, saveplot = T, mods = NA, plotloc = NA,
  pdfrows = 4, pdfcols = 2)
}
\arguments{
\item{SPRseries}{csv}

\item{saveplot}{logical. should plots be saved}

\item{plotloc}{folder to save plots; if NA, will save in location of cpue_results}
}
\description{
\code{plotF} plots vector of F_std from SS3 report file, with optional overlay of multiple models
}
\seealso{
\code{\link[r4ss]}
}
