% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInputSel.R
\name{plotInputSel}
\alias{plotInputSel}
\title{plotInputSel
\code{plotInputSel} parsing of input selectivity plots to show by-fleet, by-Sex curves as proposed in SS3 Control File}
\usage{
plotInputSel(rootdir, pattern = NA, year = 2016, seltype = c("age",
  "length")[1], plotloc = NA, lmin = 35, lmax = 350, linc = 5,
  pdfrows = 3, pdfcols = 2)
}
\arguments{
\item{moddir}{filepath where Report.sso is stored}

\item{write}{logical. should plots be saved to model directory}

\item{type}{type of file to be saved}
}
\description{
plotInputSel
\code{plotInputSel} parsing of input selectivity plots to show by-fleet, by-Sex curves as proposed in SS3 Control File
}
