% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKobe.R
\name{plotKobe}
\alias{plotKobe}
\title{plotKobe
terminal kobe plot for a singly model. check out FLR.}
\usage{
plotKobe(
  summaryoutput,
  terminal_only = c(T, F)[1],
  terminal_year = 2015,
  axes.limits = c(2, 2),
  kobe.type = c(NA, "ISC")[1],
  b.name = NA,
  f.name = NA,
  saveplot = T,
  plotloc = NA,
  doDistrib = FALSE,
  doLegend = T
)
}
\arguments{
\item{terminal_only}{Do you only want to show the final year or the whole timeseries?}

\item{terminal_year}{Either a value or vector corresponding to the terminal year in each model.}

\item{axes.limits}{x and y maximums for kobe plot, default to 2}

\item{kobe.type}{if 'ISC', will have pale orange coloring in place of orange/yellow}

\item{b.name}{the character column name of the biomass ref point; can rename to change y lab}

\item{f.name}{the character column name of the fishing ref point}

\item{saveplot}{logical, do you want to save the plot}

\item{plotloc}{where to save plot}

\item{doLegend}{logical. do you want to write a legend (not advised for huge rep sets)}

\item{rootdir}{home directory}

\item{pattern}{1st order directory character}

\item{subpattern}{2nd order directory character}
}
\description{
plotKobe
terminal kobe plot for a singly model. check out FLR.
}
