% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKobe_compare.R
\name{plotKobe_compare}
\alias{plotKobe_compare}
\title{plotKobe_compare
terminal kobe plot. check out FLR.}
\usage{
plotKobe_compare(rootdir, mq_csv = NA, b.name = NA, f.name = NA,
  pattern = NA, subpattern = NA, saveplot = T, plotloc = NA,
  doLegend = T)
}
\arguments{
\item{rootdir}{home directory}

\item{mq_csv}{optional; if you already generated a MQ csv using extractResults it will read from that, otherwise need pattern/subpattern}

\item{b.name}{the character column name of the biomass ref point; can rename to change y lab}

\item{f.name}{the character column name of the fishing ref point}

\item{pattern}{1st order directory character}

\item{subpattern}{2nd order directory character}

\item{saveplot}{logical, do you want to save the plot}

\item{plotloc}{where to save plot}

\item{doLegend}{logical. do you want to write a legend (not advised for huge rep sets)}
}
\description{
plotKobe_compare
terminal kobe plot. check out FLR.
}
