% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKobe_panel.R
\name{plotKobe_panel}
\alias{plotKobe_panel}
\title{plotKobe_panel
\code{plotKobe_panel} custom wrapper to loop through nested operating/estimation models and plot using r4ss}
\usage{
plotKobe_panel(
  rootdir,
  pattern = NA,
  plotloc = NA,
  plotrows = 1,
  plotcols = 2,
  saveplot = T
)
}
\arguments{
\item{rootdir}{master location of all models and (optional) replicates; will search for Report.sso files therein}

\item{pattern}{either a pattern to match many models or the full directory name for a single model.}

\item{plotloc}{folder to save plots; if NA, will save in location of cpue_results}

\item{plotrows}{number of rows per page of PDF}

\item{plotcols}{number of cols per page of PDF}

\item{saveplot}{logical. should plots be saved to model directory}
}
\description{
plotKobe_panel
\code{plotKobe_panel} custom wrapper to loop through nested operating/estimation models and plot using r4ss
}
