% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMnwtMatrix.R
\name{plotMnwtMatrix}
\alias{plotMnwtMatrix}
\title{plotMnwtMatrix
\code{plotMnwtMatrix} make a colored matrix showing years/sexes where model-estimated mean weight differed from input.}
\usage{
plotMnwtMatrix(summaryoutput, surveydata, fleet = 1, maxage = 30,
  fillNA = F, png = F, printfolder)
}
\arguments{
\item{summaryoutput}{generated via SS_output()}

\item{surveydata}{raw data frame -- in format from NWFSCAssess/survey}

\item{fleet}{defaults to 1; can currently only take one value at a time}

\item{maxage}{the maximum age you'd like the plot to generate to}

\item{print}{do you want the plot to save in y}
}
\description{
plotMnwtMatrix
\code{plotMnwtMatrix} make a colored matrix showing years/sexes where model-estimated mean weight differed from input.
}
