% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPacific.R
\name{plotPacific}
\alias{plotPacific}
\title{plotPacific
A ggplot basemap that is Pacific-centered. Defaults to full North Pacific}
\usage{
plotPacific(
  ew.breaks = c(seq(-120, -170, -10), 180, seq(170, 80, -10)),
  ns.breaks = seq(0, 60, 10),
  fillcol = "grey88",
  bordercol = "skyblue3",
  alpha = 1
)
}
\arguments{
\item{ew.breaks}{in 180-degree centered coordinates, provide your E/W limits and breaks}

\item{ns.breaks}{in 0-degree centered coordinates, provide your E/W limits and breaks; 0 is minimum}

\item{fillcol}{color to fill countries}

\item{bordercol}{color to outline countries}

\item{alpha}{optional transparency, defaults to 1}
}
\description{
plotPacific
A ggplot basemap that is Pacific-centered. Defaults to full North Pacific
}
