% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight_discards.R
\name{reweight_discards}
\alias{reweight_discards}
\title{reweight_discards
beta
a function to re-weight observed discard rates from WCGOP data based on pre/post catch-shares years
returns a table that can be saved/copied into a stock synthesis file. Assumes the ncs/cs dataframes
are in the traditional format as from the Observer program}
\usage{
reweight_discards(ncs, cs, fleet, fleetno = 1, month = 7, years, units,
  writeTable = F, writeloc = paste0(getwd(), "discard_rates_ss_",
  Sys.Date(), ".csv"))
}
\arguments{
\item{ncs}{a data frame of non-catch-shares observer data}

\item{cs}{a data frame of catch-shares observer data}

\item{fleet}{a string or vector of strings matching values from column gear3}

\item{fleetno}{optional; fill in any number to the "FltSvy" column of the resultant table}

\item{month}{optional; fill in a month into the month column of the resultant table}

\item{years}{optional; if you'd like to only take a subset of years from the final dataframe}

\item{units}{one of lbs or kg, if kg LBS will be converted}

\item{writeTable}{logical. if T, requires writeloc}

\item{writeloc}{a filepath to where you want the table saved.}
}
\description{
reweight_discards
beta
a function to re-weight observed discard rates from WCGOP data based on pre/post catch-shares years
returns a table that can be saved/copied into a stock synthesis file. Assumes the ncs/cs dataframes
are in the traditional format as from the Observer program
}
