% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd2rscript.R
\name{rmd2rscript}
\alias{rmd2rscript}
\title{Convert RMD to R SCRIPT. Will save a file same location with noncode blocks commented out.
 adapted from http://rstudio-pubs-static.s3.amazonaws.com/12734_0a38887f19a34d92b7311a2c9cb15022.html
  @export}
\usage{
rmd2rscript(infile)
}
\description{
Convert RMD to R SCRIPT. Will save a file same location with noncode blocks commented out.
 adapted from http://rstudio-pubs-static.s3.amazonaws.com/12734_0a38887f19a34d92b7311a2c9cb15022.html
  @export
}
