% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleIBM.R
\name{simpleIBM}
\alias{simpleIBM}
\title{simpleIBM
generalized version of IBM used for Kapur et al. growth study (Submitted to Fisheries Research). Right now this has constant M, and is not set-up to intake fishing mortality (therefore there is no selectivity). The population is initialized at equilibrium.}
\usage{
simpleIBM(
  nboot = 50,
  simu_year = 100,
  burn_yr = 20,
  Inst_M = 0.2,
  L1 = 10,
  a1 = 1,
  L2 = 100,
  a2 = 15,
  VBGF_K = 0.25,
  lw_a = 2e-04,
  lw_b = 3,
  r_mat = 0.1,
  L50 = 44,
  growth_lognormal_SD = 0.015,
  steepBH = 0.9,
  R0 = 10
)
}
\arguments{
\item{nboot}{the number of unique simulations of length simu_year you'd like to generate}

\item{simu_year}{number of years to conduct simulation}

\item{Inst_M}{natural mortality (single value applied to all ages)}

\item{L1}{start length}

\item{a1}{start age}

\item{L2}{terminal length}

\item{a2}{terminal age}

\item{VBGF_K}{K as in the von bertalanffy growth function}

\item{lw_a}{coefficient of length-weight exponential equation}

\item{lw_b}{exponent of length-weight exponential equation}

\item{r_mat}{slope of maturity ogive}

\item{L50}{length at 50 maturity}

\item{steepBH}{Beverton-holt steepness}

\item{burn_year}{can be zero; number of year's youd like to remove from the beginning of the simulation}
}
\description{
simpleIBM
generalized version of IBM used for Kapur et al. growth study (Submitted to Fisheries Research). Right now this has constant M, and is not set-up to intake fishing mortality (therefore there is no selectivity). The population is initialized at equilibrium.
}
