% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_autoForecast.R
\name{SS_autoForecast}
\alias{SS_autoForecast}
\title{SS_autoForecast}
\usage{
SS_autoForecast(rootdir, basedir, state = state, statesex = 1,
  statevals = 0.05, catch_proportions = c(0.5, 0.08426184, 0.4157382),
  forecast_start = 2021, forecast_end = 2031,
  fixed_catches = catch_projections[1:4, 5:7],
  Flimitfraction = catch_projections$PSTAR_0.45[catch_projections$YEAR >
  2020])
}
\arguments{
\item{rootdir}{somewhere you'd like several models in folders created, eg forecast202X}

\item{basedir}{directory with executed base-case model -- right now, assumed to be inside rootdir}

\item{state}{one of low/med/high -- only works for natural mortality at present}

\item{statesex}{0, 1 or 2 for female, male or both}

\item{statevals}{a dataframe with columns corresponding to state (labeled low/medium/high), and optionally rows corresponding to Female and Male values}

\item{catch_proportions}{a single or vector of values denoting the allocation proportiosn for each fleet in order matching .dat file; assuming F_relative opt 2}

\item{forecast_start}{the first year to forecast; assume inputs before this}

\item{forecast_end}{last year to forecast}

\item{fixed_catches}{a matrix of input fixed catches from the end of the original model to forecast_start}

\item{Flimitfraction}{a value or fector of same length as forecast period with P* values corresponding to a given year}
}
\description{
#' \code{SS_autoForecast} beta version of tool to automate catch-only update iterations and development of decision tables
hopefully port to r4ss when ready
}
