% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_compare.R
\name{ss_compare}
\alias{ss_compare}
\title{ss_compare
custom wrapper to loop through nested operating/estimation models and plot using r4ss}
\usage{
ss_compare(rootdir, subplots = 1:20, plotloc = NA, pattern = "Rep",
  llabels = c(paste0("OMR", 1:2), paste0("EMR", 1:2)), likeCSV = T,
  likeLabel = c("Surv_like", "SizeFreq_like:_2"),
  lambdaLabel = c("Surv_lambda", "SizeFreq_lambda:_2"),
  fishery = "Deep7", fleetIdx = c(2, 3), dolegend = T)
}
\arguments{
\item{rootdir}{master location of all models and (optional) replicates; will search for Report.sso files therein}

\item{pattern}{a string which identifies the first-order folders, e.g "EM" "OM" -- applicable to grep().}

\item{llabels}{logical. If T, will write a CSV with results.}

\item{subpattern}{can be NA; a secondary string by which models will be grouped, e.g. "Replicate"}
}
\description{
ss_compare
custom wrapper to loop through nested operating/estimation models and plot using r4ss
}
