% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_black.R
\name{theme_black}
\alias{theme_black}
\title{custom GGPLOT2 theme for insertion
adapted from https://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2/}
\usage{
theme_black(base_size = 12, base_family = "")
}
\description{
custom GGPLOT2 theme for insertion
adapted from https://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2/
}
