% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mk.R
\name{theme_mk}
\alias{theme_mk}
\title{custom GGPLOT2 theme for insertion
adapted from https://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2/}
\usage{
theme_mk(base_size = 12, base_family = "")
}
\description{
custom GGPLOT2 theme for insertion
adapted from https://jonlefcheck.net/2013/03/11/black-theme-for-ggplot2-2/
}
