% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorTables.R
\name{errorTables}
\alias{errorTables}
\title{errorTables}
\usage{
errorTables(refList, write = T)
}
\arguments{
\item{write}{logical. should plots be saved to model directory}

\item{moddir}{filepath where Report.sso is stored}

\item{type}{type of file to be saved}
}
\description{
\code{plotInputSel} parsing of input selectivity plots to show by-fleet, by-sex curves as proposed in SS3 Control File
}
\seealso{
\code{\link[r4ss]}
}
