% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{extract_results
\code{extract_results} custom wrapper of to extract reference points nested SS3 simulations}
\usage{
extractResults(
  rootdir,
  terminal_year = 2015,
  suffix = NA,
  pattern = NA,
  subpattern = NA,
  writeTables = T
)
}
\arguments{
\item{rootdir}{root filepath where all subdirectories containing Report.sso are stored}

\item{terminal_year}{last year for reference pt extraction}

\item{suffix}{optional tag appended to end of CSV filenames; defaults to subpattern or pattern}

\item{pattern}{a string that specifically matches all directories with report files of interest.}

\item{subpattern}{type of file to be saved}

\item{writeTables}{logical. Should a csv of results be saved?}

\item{FleetName}{optional vector of fleet names for which data should be extracted; all means all}
}
\description{
extract_results
\code{extract_results} custom wrapper of to extract reference points nested SS3 simulations
}
