% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{extract_results
\code{extract_results} custom wrapper of to extract reference points nested SS3 simulations}
\usage{
extractResults(rootdir, terminal_year = 2015, pattern = NA,
  subpattern = NA, writeTables = T, FleetName = c("S4_JPN_SS",
  "S7_JPN_GEO", "All")[3])
}
\arguments{
\item{rootdir}{root filepath where all subdirectories containing Report.sso are stored}

\item{pattern}{a string that specifically matches all directories with report files of interest.}

\item{subpattern}{type of file to be saved}

\item{writeTables}{logical. Should a csv of results be saved?}

\item{FleetName}{optional vector of fleet names for which data should be extracted; all means all}
}
\description{
extract_results
\code{extract_results} custom wrapper of to extract reference points nested SS3 simulations
}
