% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestidx.R
\name{nestidx}
\alias{nestidx}
\title{Nested loop index}
\usage{
nestidx(vals = list(a, b, b.vector, c = 0, c.vector = 0))
}
\arguments{
\item{vals}{a list of up to length three with the current indices and length of each index vector}
}
\description{
\code{nestidx} returns the index value for a matrix populated by nested loops.
}
\examples{
a.vector = seq(1:4)
b.vector = seq(7:100)
c.vector = seq(12:14)
#' mk <- matrix(NA)
for(a in 1:length(a.vector)){
for(b in 1:length(b.vector)){
 for(c in 1:length(c.vector)){
    idx <- nestidx(vals = list(a,b,b.vector,c,c.vector))
      cat(idx,"\\n")
      mk[idx] <- a*b
   }
 }
}
}
