% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKobe_compare.R
\name{plotKobe_compare}
\alias{plotKobe_compare}
\title{plotKobe_compare
terminal kobe plot. check out FLR.}
\usage{
plotKobe_compare(rootdir, axes.limits = c(2, 2),
  terminal_year = c(2017, 2017, 2016, 2017), kobe.type = c(NA,
  "ISC")[1], mq_csv = NA, b.name = NA, f.name = NA, pattern = NA,
  subpattern = NA, saveplot = T, plotloc = NA, doDistrib = FALSE,
  doLegend = T)
}
\arguments{
\item{rootdir}{home directory}

\item{axes.limits}{x and y maximums for kobe plot, default to 2}

\item{terminal_year}{Either a value or vector corresponding to the terminal year in each model.}

\item{kobe.type}{if 'ISC', will have pale orange coloring in place of orange/yellow}

\item{mq_csv}{optional; if you already generated a MQ csv using extractResults it will read from that, otherwise need pattern/subpattern}

\item{b.name}{the character column name of the biomass ref point; can rename to change y lab}

\item{f.name}{the character column name of the fishing ref point}

\item{pattern}{1st order directory character}

\item{subpattern}{2nd order directory character}

\item{saveplot}{logical, do you want to save the plot}

\item{plotloc}{where to save plot}

\item{doDistrib}{logical. Do you want to have FLR style distributions on outer margins? Only works if models run with hessian.}

\item{doLegend}{logical. do you want to write a legend (not advised for huge rep sets)}
}
\description{
plotKobe_compare
terminal kobe plot. check out FLR.
}
